scripts/multi-agent/requirements.txt
openai>=1.82.0
openai-agents>=0.1.0
python-dotenv>=1.0.0
